requirements.txt
fastapi==0.115.5
uvicorn[standard]
pydantic[email]
pydantic-settings==2.6.1
async-sqlalchemy==1.0.0
asyncpg==0.30.0